chore: slipt Products tests into "unit" and "functinal" and reorganize it into subfolders
